// src/types/DrawerLayer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DrawerAdjust } from "./DrawerAdjust";
import type { DrawerBlendMode } from "./DrawerBlendMode";
import type { DrawerObject } from "./DrawerObject";
import type { DrawerTransform } from "./DrawerTransform";

export type DrawerLayer = { id: string, name: string, visible: boolean, opacity: number, objects: Array<DrawerObject>, blend: DrawerBlendMode, transform: DrawerTransform | null, adjust: DrawerAdjust | null, };